requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
seaborn
imbalanced-learn
joblib
requests
fastapi
uvicorn
streamlit
google-generativeai
pydantic
python-multipart
jupyter
pytest
